google-dataflow-dota-2-gsi-streaming/statistics_rest_api/requirements.txt
google-cloud-datastore==1.7.3